feat(main): Create canvas module, apply to ptest
The original particle test was rife with bugs and failure. Step one of
correcting this was to separate main canvas control code out to its own
module. Step two was to incorporate this module into a new ptest design
with a more incremental approach to make sure each piece of code was
functioning correctly. Yet to do is refine the particle test into a
possible site background animation.